Been a while...
- Placed related files in separated subfolders for organization sake
- Changed header format from .h to .hpp

After doing some school projects a few months back and finally getting over my lack of motivation I decided to pick up this project and see if I can finish it.

+++ src/Gui/gui.cpp
@@ -5,8 +5,8 @@ extern "C"{
     #include"raygui.h"
 }
 
-#include"brush.h"
-#include"gui.h"
+#include"Brush/brush.hpp"
+#include"gui.hpp"
 
 bool drawGui()
 {
@@ -15,7 +15,7 @@ bool drawGui()
         GuiSlider((Rectangle){0, 192, 192, 20}, nullptr, nullptr, &(Brush::brushSize), 0, 20);
     };
 
-    Panel panel1(content1, (Rectangle){0, 0, 200, GetScreenHeight()});
+    Panel panel1(content1, (Rectangle){0, 0, 200, (float)GetScreenHeight()});
     panel1._draw();
 
     return panel1.IsOverPanel(GetMousePosition());
